+++ scripts/package_macos_app.sh
@@ -58,9 +58,9 @@ cat > "$CONTENTS_DIR/Info.plist" <<'PLIST'
   <key>CFBundlePackageType</key>
   <string>APPL</string>
   <key>CFBundleShortVersionString</key>
-  <string>0.0.00098</string>
+  <string>0.0.00099</string>
   <key>CFBundleVersion</key>
-  <string>98</string>
+  <string>99</string>
   <key>LSMinimumSystemVersion</key>
   <string>14.0</string>
   <key>NSHighResolutionCapable</key>